+++ tests/files/src/0468_unparseable_param.php
@@ -1,0 +1,10 @@
+<?php
+
+/**
+ * @param int? $x
+ * @param int
+ * @return int?
+ */
+function exampleSuspiciousComment($x, $y) {
+    return $x;
+}